Final GA munge (and KY test)
