+++ src/ini_files_for_results/Georgia/ga20g_20201120_1237.ini
@@ -1,0 +1,12 @@
+[election_data_analysis]
+results_file=Georgia/GA_detail_20201120_1237.xml
+jurisdiction_directory=Georgia
+munger_name=ga_xml
+top_reporting_unit=Georgia
+election=2020 General
+results_short_name=ga20g_1120_1237
+results_download_date=2020-11-20
+results_source=https://results.enr.clarityelections.com//GA//105369/
+results_note=
+aux_data_dir=
+is_preliminary=False
